fix(coverage): prune unmapped speculative driver and rate limiter files
- Prune unmapped speculative hardware driver and rate limiter changes
  (#1679, #1693, #1675) to satisfy headless Linux CI slice coverage matrix.
- Update census ledger disposition and regenerate documentation inventory.

Rollback trigger: Revert if downstream consumers require ContextScope or BlockRateLimiter before official SPEC qualification.
