+++ crates/ramshared-cuda/src/ffi.rs
@@ -27,8 +27,6 @@ pub type FnDeviceGetName = unsafe extern "C" fn(*mut c_char, c_int, CuDevice) ->
 pub type FnCtxCreate = unsafe extern "C" fn(*mut CuContext, c_uint, CuDevice) -> CuResult;
 pub type FnCtxDestroy = unsafe extern "C" fn(CuContext) -> CuResult;
 pub type FnCtxSynchronize = unsafe extern "C" fn() -> CuResult;
-pub type FnCtxPushCurrent = unsafe extern "C" fn(CuContext) -> CuResult;
-pub type FnCtxPopCurrent = unsafe extern "C" fn(*mut CuContext) -> CuResult;
 pub type FnMemAlloc = unsafe extern "C" fn(*mut CuDevicePtr, usize) -> CuResult;
 pub type FnMemFree = unsafe extern "C" fn(CuDevicePtr) -> CuResult;
 pub type FnMemcpyHtoD = unsafe extern "C" fn(CuDevicePtr, *const c_void, usize) -> CuResult;
@@ -46,8 +44,6 @@ pub struct Syms {
     pub ctx_create: FnCtxCreate,
     pub ctx_destroy: FnCtxDestroy,
     pub ctx_synchronize: FnCtxSynchronize,
-    pub ctx_push_current: FnCtxPushCurrent,
-    pub ctx_pop_current: FnCtxPopCurrent,
     pub mem_alloc: FnMemAlloc,
     pub mem_free: FnMemFree,
     pub memcpy_htod: FnMemcpyHtoD,
