+++ crates/ramshared-cuda/src/lib.rs
@@ -30,13 +30,11 @@ mod loader_win;
 #[cfg(windows)]
 use loader_win as loader;
 
-pub mod context_scope;
 mod driver;
 mod ffi;
 pub mod probe;
 mod vram_impl; // impl VramProvider/VramMemory for CUDA types (RF-G1)
 
-pub use context_scope::ContextScope;
 pub use driver::{Context, Cuda, CudaError, Device, DeviceMem};
 pub use probe::{PROBE_PATTERN_LEN, ProbePlanError, pattern_for_offset, plan_probe_offsets};
